feat(v3.11/4903): user export list add project info
